application_layer/app/core/services/ai_service.py
# core/services/ai_service.py

from typing import List, Optional

from core.config.settings import settings
from google.cloud import firestore


class AIService:
    def __init__(self):
        self.db = firestore.Client()

    def get_product_recommendations(self, user_id: str) -> List[dict]:
        """Get product recommendations for a user"""
        try:
            # In a real implementation, this would use a trained ML model
            # For now, return popular products as a placeholder
            query = self.db.collection(f"{settings.ENVIR}_robots").order_by("ratings.average", direction=firestore.Query.DESCENDING).limit(5)
            
            recommendations = []
            for doc in query.stream():
                recommendations.append({"product_id": doc.id, **doc.to_dict()})
            
            return recommendations
        except Exception as exc:
            print(f"Error getting product recommendations: {exc}")
            return []

    def detect_fraud(self, transaction_data: dict) -> dict:
        """Detect fraudulent transactions"""
        try:
            # In a real implementation, this would use a fraud detection model
            # For now, use a simple rule-based approach
            is_fraudulent = False
            if transaction_data.get("amount", 0) > 10000:
                is_fraudulent = True
            
            return {"is_fraudulent": is_fraudulent, "score": 0.9 if is_fraudulent else 0.1}
        except Exception as exc:
            print(f"Error detecting fraud: {exc}")
            return {"is_fraudulent": False, "score": 0}

    def monitor_compliance(self, data: dict) -> dict:
        """Monitor for compliance with platform policies"""
        try:
            # In a real implementation, this would use NLP to analyze text data
            # For now, use a simple keyword check
            is_compliant = "prohibited" not in data.get("description", "").lower()
            
            return {"is_compliant": is_compliant}
        except Exception as exc:
            print(f"Error monitoring compliance: {exc}")
            return {"is_compliant": True}
