application_layer/requirements.txt
fastapi==0.95.1
uvicorn
pydantic
requests==2.28.2
debugpy # Required for debugging.
google-cloud-storage
google-cloud-secret-manager
google-cloud-firestore
web3
bleach
Jinja2==3.1.2
python-jose[cryptography]
passlib[bcrypt]
